Adding a perf_timed decorator which times every call of the function
also an imgui renderer for easy viewing of the timings.

due to current limitations it seems like function names are as explicit as I would have hoped so gonna need to improve that in the future. For now just decorate the functions you want tracked (maybe we should have a name argument?)